added access control to the controllers

+++ app/Http/Controllers/ReportController.php
@@ -7,8 +7,13 @@ use App\Http\Requests;
 use App\User;
 use App\Report;
 use App\Http\Requests\ReportRequest;
+use Illuminate\Support\Facades\Gate;
 
 class ReportController extends Controller {
+
+    private function showDashboard() {
+        return redirect()->route('dashboard');
+    }
 
     /**
      * Display a listing of the resource.
@@ -25,9 +30,14 @@ class ReportController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function create() {
-        $title    = "Create Report";
-        $patients = User::patients()->get();
-        return view("login.create-report", compact("title", 'patients'));
+        if (Gate::denies('is_admin')) {
+            return $this->showDashboard();
+        }
+        else {
+            $title    = "Create Report";
+            $patients = User::patients()->get();
+            return view("login.create-report", compact("title", 'patients'));
+        }
     }
 
     /**
@@ -37,17 +47,19 @@ class ReportController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function store(ReportRequest $request) {
-        $report                   = new Report();
-        $report->user_id          = $request->input('patient');
-        $report->test_date        = $request->input('testdate');
-        $report->testing_lab      = $request->input('testedby');
-        $report->case_number      = str_random(12);
-        $report->report_name      = $request->input('reportname');
-        $report->patient_history  = $request->input('history');
-        $report->description      = $request->input('reportdetails');
-        $report->addition_details = $request->input("additionaldetails");
-        $report->save();
-        return redirect()->route('dashboard');
+        if (Gate::allows('is_admin')) {
+            $report                   = new Report();
+            $report->user_id          = $request->input('patient');
+            $report->test_date        = $request->input('testdate');
+            $report->testing_lab      = $request->input('testedby');
+            $report->case_number      = str_random(12);
+            $report->report_name      = $request->input('reportname');
+            $report->patient_history  = $request->input('history');
+            $report->description      = $request->input('reportdetails');
+            $report->addition_details = $request->input("additionaldetails");
+            $report->save();
+        }
+        return $this->showDashboard();
     }
 
     /**
@@ -57,9 +69,14 @@ class ReportController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function show($report) {
-        $title = "Report Details";
-        $report = Report::findOrFail($report);
-        return view("login.report-operator", compact("report","title"));
+        if (Gate::allows('view_reports', $report)) {
+            $title  = "Report Details";
+            $report = Report::findOrFail($report);
+            return view("login.report-operator", compact("report", "title"));
+        }
+        else {
+            return $this->showDashboard();
+        }
     }
 
     /**
@@ -69,10 +86,15 @@ class ReportController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function edit($report) {
-        $report   = Report::findOrFail($report);
-        $patients = User::patients()->get();
-        $title    = "Edit Report";
-        return view("login.edit-report", compact("patients", "title", "report"));
+        if (Gate::allows("is_admin")) {
+            $report   = Report::findOrFail($report);
+            $patients = User::patients()->get();
+            $title    = "Edit Report";
+            return view("login.edit-report", compact("patients", "title", "report"));
+        }
+        else {
+            return $this->showDashboard();
+        }
     }
 
     /**
@@ -83,16 +105,18 @@ class ReportController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function update(ReportRequest $request, $report) {
-        $report = Report::findOrFail($report);
-        $report->user_id          = $request->input('patient');
-        $report->test_date        = $request->input('testdate');
-        $report->testing_lab      = $request->input('testedby');
-        $report->report_name      = $request->input('reportname');
-        $report->patient_history  = $request->input('history');
-        $report->description      = $request->input('reportdetails');
-        $report->addition_details = $request->input("additionaldetails");
-        $report->save();
-        return redirect()->route('dashboard');
+        if (Gate::allows("is_admin")) {
+            $report                   = Report::findOrFail($report);
+            $report->user_id          = $request->input('patient');
+            $report->test_date        = $request->input('testdate');
+            $report->testing_lab      = $request->input('testedby');
+            $report->report_name      = $request->input('reportname');
+            $report->patient_history  = $request->input('history');
+            $report->description      = $request->input('reportdetails');
+            $report->addition_details = $request->input("additionaldetails");
+            $report->save();
+        }
+        return $this->showDashboard();
     }
 
     /**
@@ -102,8 +126,10 @@ class ReportController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function destroy($report) {
-        Report::destroy($report);
-        return redirect()->route('dashboard');
+        if (Gate::allows("is_admin")) {
+            Report::destroy($report);
+        }
+        return $this->showDashboard();
     }
 
 }
